#143 Wrong render on null value - fixed
