+++ app/views/measurements/show.html.erb
@@ -14,7 +14,7 @@
   <dd><%= @measurement.id %></dd>
 
   <dt><%= t('activerecord.attributes.measurement.captured_at') %></dt>
-  <dd><%= @measurement.captured_at_formatted %></dd>
+  <dd><%= @measurement.captured_at.nil? ? "null" : @measurement.captured_at_formatted %></dd>
 
   <%- if @measurement.user.present? -%>
     <dt><%= t('activerecord.attributes.measurement.submitted_by') %></dt>
